requirements.txt
anyio==4.9.0
certifi==2025.7.14
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
python-dotenv==1.1.1
python-telegram-bot>=22,<23
sniffio==1.3.1
python-Levenshtein>=0.27.0
pyairtable==2.3.3
